+++ packages/unified/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.18](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/unified@1.0.17...@amplitude/unified@1.0.18) (2026-04-16)
+
+**Note:** Version bump only for package @amplitude/unified
+
+
+
+
+
 ## [1.0.17](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/unified@1.0.16...@amplitude/unified@1.0.17) (2026-04-14)